backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
Flask-Marshmallow==0.15.0
marshmallow==3.26.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
mysql-connector-python==8.1.0
redis==5.0.0
celery==5.3.1
pytest==7.4.2
pytest-flask==1.2.0
faker==19.6.2
openai>=1.99.0
stripe>=7.0.0
requests>=2.31.0
cryptography>=41.0.0
PyJWT>=2.8.0
pytz==2023.3
APScheduler==3.10.4
Flask-SocketIO==5.3.6
pydantic==2.5.0